+++ modules/acme/default.nix
@@ -24,11 +24,12 @@ in {
   security.acme = lib.mkIf prod {
     acceptTerms = true;
 
-    certs = merge (builtins.map (tld: {
-      "${tld}" = {
-        domain = "*.${tld}";
+    certs = merge (builtins.map (domain: {
+      "${domain}" = {
+        inherit domain;
         dnsProvider = "cloudflare";
         environmentFile = "${config.age.secrets.acme-environment-file.path}";
+        extraDomainNames = "*.${domain}";
         reloadServices = [ "nginx" ];
       };
     }) config.services.nginx.domains);