+++ hosts/gastly/default.nix
@@ -1,5 +1,4 @@
 { config, pkgs, lib, self, ... }:
-
 let
   inherit (self) common;
   inherit (self.common.home-manager-module-sets) hyprland-waybar-desktop;
@@ -26,9 +25,11 @@ in {
     greetd
     grub
     hyprland
+    i18n
     lorri
     nix
     openssh
+    pipewire
     steam
     systemd-networkd
     time
@@ -122,6 +123,9 @@ in {
   };
 
   powerManagement.cpuFreqGovernor = lib.mkDefault "powersave";
+
+  swapDevices =
+    [{ device = "/dev/disk/by-uuid/907b7556-218a-4516-ae2b-0b310d6b0b19"; }];
 
   system.stateVersion = "22.11";
 }