+++ packages/python/docxtpl/default.nix
@@ -19,14 +19,7 @@ let
 
 in buildPythonPackage {
   inherit pname version meta;
-  propagatedBuildInputs = [
-    six
-    (python-docx.overridePythonAttrs
-      (_: { disabledTests = [ "DescribeParseXml" ]; }))
-    jinja2
-    lxml
-    docxcompose
-  ];
+  propagatedBuildInputs = [ six python-docx jinja2 lxml docxcompose ];
 
   src = fetchPypi {
     inherit pname version;